+++ src/_dot_product/c3_14/dprod.cpp
@@ -13,8 +13,8 @@ sundiazontai metaksu tous, me ton combiner operator.
 /* COMMENT
  * Critical does not compile with simd. It needs simd ordered
  */
-float dprod(size_t num, float *a, float *b) {
-	float res = 0.0;
+double dprod(size_t num, double *a, double *b) {
+	double res = 0.0;
 #pragma omp target teams map(res) map(tofrom: a[0:num], b[0:num])
 #pragma omp distribute parallel for simd 
 		for (size_t i = 0; i < num; ++i) {
