+++ src/_dot_product/c3_7_1/dprod.cpp
@@ -13,8 +13,8 @@ sundiazontai metaksu tous, me ton combiner operator.
 /* COMMENT
  * einai i parallagi tis c3_7 pou omos vgazei sosta apotelesmata
  */
-float dprod(size_t num, float *a, float *b) {
-	float res = 0.0;
+double dprod(size_t num, double *a, double *b) {
+	double res = 0.0;
 #pragma omp target  defaultmap(tofrom: scalar) map(to: a[0:num], b[0:num])
 #pragma omp teams reduction(+ : res)
 #pragma omp distribute parallel for reduction(+ : res)
